requirements.txt
attrs==22.1.0
black==22.8.0
click==8.1.3
colorama==0.4.5
Flask==2.2.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.0
greenlet==1.1.3
gunicorn==20.1.0
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy==0.971
mypy-extensions==0.4.3
mysql-connector-python==8.0.30
mysqlclient==2.1.1
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.1.3
six==1.16.0
SQLAlchemy==1.4.41
tomli==2.0.1
typing_extensions==4.3.0
Werkzeug==2.2.1
